+++ workflow/src/main/java/de/esempe/workflow/boundary/db/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import de.esempe.workflow.domain.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>, AdditionalUserRepository
+public interface UserRepository extends JpaRepository<User, Long>, AdditionalUserRepository
 {
 	// "automatische" Query "byName" durch Spring
 	List<User> findByLastname(String lastname);
